requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pymongo==4.5.0
motor==3.2.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
selenium==4.12.0
webdriver-manager==4.0.0
httpx==0.24.1
python-multipart==0.0.6
aiohttp==3.8.5
aiofiles==23.2.1
redis==4.6.0
pytest==7.4.2
pytest-asyncio==0.21.1
google-generativeai==0.3.1 